Disable airline scrollbar extension
An extension was added to the airline plugin to represent progress
through a file via a scrollbar graphic. However, this took up a lot of
space on the status line, shortening the amount available for the
current file name. It is also pointless when the line number and
percentage is already shown in the status line.